results are now fetched one by one on user's demand (using AJAX)
Only one result is fetched when user clicks button for generating valid timetables, and for getting each other result, they should click button '>' what will cause that the new result will be fetched and displayed (or user will get a message that there are no any other results). In Prolog program can now be set as fact which timetable was previously generated, so in order to find another valid timetable (if there is any), program doesn't need again to iterate over all combinations that were checked in the previous request-session.
Result of this approach is that, in case there is a lot of solutions (or combinations to check where some of them match set criteria), user doesn't need to wait until all combinations are found, but the found result is immediately returned.
Summary: this solution is good if user just wants to get some results and as soon as possible, but it can take user very long time if there is a lot of results and user wants to go through all (or a lot) of them