results are now fetched continuously asynchronously (using WebSockets)
When user presses button for generating valid timetables, server doesn't return any timetables, but only port number of listening socket on Daemon process that executes PHP script - user's browser connects to that socket using WebSockets and receives all the data through it in real-time (currently every second).
Result of this approach is that, in case there is a lot of solutions (or combinations to check where some of them match set criteria), user doesn't need to wait until all combinations are found, but the found result is immediately returned. Only disadvantage might be that WebSockets are not supported on older web-browsers.
Summary: this solution is good even if user just wants to get some results and as soon as possible, but it is also appropriate if there is a lot of results and user wants to go through all (or a lot) of them